backend/requirements.txt
django==4.2.10
djangorestframework==3.14.0
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.1
django-cors-headers==4.3.1
gunicorn==21.2.0
Pillow==10.2.0
